Rename files to follow Kolibri naming convention
