+++ kolibri/plugins/style_guide/assets/src/views/content/components/simple-modals.vue
